Don't destroy background image after iteration one
imagedestroy($background) is called after the first iteration but
possibly used in further iteration of user solutions.

Maybe tell your workmates that while keeping your question type plugins
up-to-date, you aren't maintaining ILIAS and Opencast anymore.
The impression I always get when talking to Rektorat or LLZ or even my
boss is that maintaining ILIAS is a matter of LLZ or you.